+++ packages/components-angular/src/app/app.module.ts
@@ -1,9 +1,10 @@
-import { BrowserModule } from '@angular/platform-browser';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { BrowserModule } from '@angular/platform-browser';
+import { IMaskModule } from 'angular-imask';
 import { PorscheDesignSystemModule } from '@porsche-design-system/components-angular';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
-import { IMaskModule } from 'angular-imask';
 import * as fromComponents from './components';
 import * as fromPages from './pages';
 import * as fromExamples from './examples';
@@ -20,10 +21,11 @@ import * as fromStyles from './styles';
   ],
   imports: [
     BrowserModule,
+    FormsModule,
     AppRoutingModule,
     IMaskModule,
-    PorscheDesignSystemModule.load({ prefix: '', cdn: 'auto' }),
-    PorscheDesignSystemModule.load({ prefix: 'my-prefix', cdn: 'auto' }),
+    PorscheDesignSystemModule.load({ prefix: '', cdn: 'auto' }), // first configured theme wins or fallback 'light'
+    PorscheDesignSystemModule.load({ prefix: 'my-prefix', cdn: 'auto', theme: 'dark' }),
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
   bootstrap: [AppComponent],